feat: meld LineInterface into BaseModel
